ref config websocket
